pyproject.toml
[project]
name = "stock-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "aiohttp>=3.8.0",
    "akshare>=1.17.83",
    "colorlog==6.8.2",
    "dashscope",
    "gnews==0.4.2",
    "langchain==1.0.2",
    "langchain-community==0.4",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai==1.0.1",
    "logger",
    "matplotlib==3.10.6",
    "mcp[cli]>=0.9.0",
    "numpy==1.26.4",
    "pandas==2.2.0",
    "pydantic",
    "python-dateutil==2.9.0",
    "python-dotenv==1.0.1",
    "qwen-agent[gui,mcp]>=0.0.30",
    "requests>=2.32.5",
    "serpapi==0.1.5",
    "yfinance>=0.2.36",
]
